feat: auth_middleware invalidates old sessions after password change
Set correct type i64 for claims exp, iat. Add db access to auth middleware. Use the updated_at from the user table to validate tokens. Reject all tokens whose issue timestamp is older then user's updated_at timestamp. Provide parsing for updated_at from a user table to unix timestamp.

+++ api/src/users.rs
@@ -1,3 +1,5 @@
+use std::ops::Add;
+
 use axum::{Json, http::StatusCode};
 use axum_extra::extract::cookie::{Cookie, CookieJar};
 use jsonwebtoken::{EncodingKey, Header, encode};
@@ -66,8 +68,8 @@ fn create_auth_cookie(secret: &str) -> Result<Cookie<'static>, StatusCode> {
 
     let claims = auth::Claims {
         sub: "admin".into(),
-        exp: exp.unix_timestamp() as usize,
-        iat: now.unix_timestamp() as usize,
+        exp: exp.unix_timestamp(),
+        iat: now.unix_timestamp(),
     };
 
     let token = encode(
